Conflicts:
	src/uComponents.DataTypes/ImagePoint/ImagePointDataEditor.cs

+++ src/uComponents.DataTypes/ImagePoint/ImagePointDataEditor.cs
@@ -353,12 +353,12 @@ namespace uComponents.DataTypes.ImagePoint
             }
             else if (this.options.Width > 0 && this.options.Height == 0)
             {
-                // width set, so calulate height
+                // width set, so calculate height
                 width = this.options.Width;
 
                 if (imageSize.Height > 0)
                 {
-                    height = (int)(imageSize.Height / ((decimal)(imageSize.Width / this.options.Width)));
+                    height = (int)(imageSize.Height / ((decimal)imageSize.Width / (decimal)this.options.Width));
                 }
             }
             else if (this.options.Width == 0 && this.options.Height > 0)
@@ -366,7 +366,7 @@ namespace uComponents.DataTypes.ImagePoint
                 // height set, so calculate width
                 if (imageSize.Width > 0)
                 {
-                    width = (int)(imageSize.Width / ((decimal)(imageSize.Height / this.options.Height)));
+                    width = (int)(imageSize.Width / ((decimal)imageSize.Height / (decimal)this.options.Height));
                 }
 
                 height = this.options.Height;
